refactor: 코드리뷰 내용 반영
- AuthErrorCode에 @AllArgsConstructor 적용
- auth 기능 global 패키지로 이동
