+++ BE/promate/src/main/java/org/example/promate/domain/user/service/UserService.java
@@ -7,11 +7,8 @@ import org.example.promate.domain.user.entity.User;
 import org.example.promate.domain.user.exception.UserErrorCode;
 import org.example.promate.domain.user.exception.UserException;
 import org.example.promate.domain.user.repository.UserRepository;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 
 @Service
 @RequiredArgsConstructor
